[EDIT] to int

+++ src/Reports/models.py
@@ -42,7 +42,7 @@ class SkanyReport(models.Model):
     zlecenie = models.CharField(max_length=50, blank=True, null=True)
     violation_found = models.BooleanField(blank=True, null=True)
     execution_date = models.DateTimeField(blank=True, null=True)
-    operation_time = models.FloatField(blank=True, null=True)
+    operation_time = models.BigIntegerField(blank=True, null=True)
 
     def __str__(self):
         return f"{self.id}"
